requirements.txt
pyminifier==2.1
GitPython==2.1.10
pathlib2==2.3.2
metayaml==0.22
pluginbase==0.7